requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
discord.py==1.7.3
dj-database-url==0.5.0
Django==4.0.4
django-heroku==0.3.1
filelock==3.6.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.5.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.8.0
multidict==6.0.2
parsel==1.6.0
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
python-decouple==3.6
queuelib==1.6.2
requests==2.27.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tldextract==3.2.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
w3lib==1.22.0
whitenoise==6.2.0
yarl==1.7.2
zope.interface==5.4.0
